Background power parameter relocated acording to its real functionality
In the different backgrounds types, the power parameter was shown only when IBL was enabled. However in most background types the power parameter affects the background rendering even if IBL is off. So, in those cases I've relocated the power parameter so it's always accessible even if IBL is off.

Changes to be committed:
	modified:   ui/properties_yaf_world.py