New parameter - cast shadows
As requested in http://www.yafaray.org/node/665 a new Material parameter has been added to enable/disable casting shadows. The default value is to be enabled, as that's the expected behavior for all objects in a scene. However, if certain objects need to be created shadowless (like in some architectural images), then shadow casting can be disabled in a per-material basis.

 Changes to be committed:
	modified:   io/yaf_material.py
	modified:   prop/yaf_material.py
	modified:   ui/properties_yaf_material.py