Fix for Camera panels appearing in LuxRender
When LuxRender and YafaRay are enabled in Blender, sometimes the YafaRay camera panels appear among the LuxRender panels.

I believe that's because the property COMPAT_ENGINES is set to the base class which is also set by LuxRender. In my opinion we should set this property individually for each of the YafaRay classes and not relying in the base class.

I also believe we should review all YafaRay UI python files to make sure this kind of undesired interference with other renderers does not happen with other YafaRay classes, maybe also putting this property to each YafaRay individual class?

 Changes to be committed:
	modified:   ui/properties_yaf_camera.py