Blend materials can be selected with drop-down menus
Changes to be able to select the blend materials using drop down menus, in a similar way as it was done in the past with the old (buggy) "enum" parameters. This solution combines the reliability of using string parameters to select the blend material components with the ease of use of the dropdown menus for material selection.

These changes have been adapted from povmaniaco's The Bounty renderer. Thanks, povmaniaco! :-)

 Changes to be committed:
	modified:   ui/properties_yaf_material.py