New texture interpolation parameter
As requested in the forum http://www.yafaray.org/community/forum/viewtopic.php?f=22&p=31289#p31289

It's not currently possible in the exporter to select the image texture interpolation type, but in the Core code we can actually select "none", "bilinear" (default) or "bicubic". So, I've added this parameter to the Exporter as well, in the Image Sampling section.
	modified:   io/yaf_texture.py
	modified:   prop/yaf_texture.py
	modified:   ui/properties_yaf_texture.py