Add error message if exception during tile rendering
Until now, if for some reason there was a problem with the tile rendering in the exporter, due to errors in the exporter code or (more likely) unexpected changes in the Blender Rendering API, there were no error messages at all, but the render was totally black, making troubleshooting much more difficult for users and developers.

I've added code that shows an error message if there is an exception while rendering the tiles and the cause of the exception. I hope that with this we will be able to better react to changes in Blender API, etc, that affect the Rendering interface.

While implementing this, I discovered there were exceptions in the Material Preview (didn't seem to affect the preview itself, though), so I've used the oportunity to fix that as well.

Changes to be committed:
modified: io/yaf_export.py