Fix for material preview not correctly updated
I can see that after updating certain properties as "diffuse_color" the material preview window is refreshed, but when updating non-blender properties like glossy_color, the material preview window is not updated.

Not sure if this is a Blender API problem or a YafaRay problem, but I'm starting to suspect Blender could be the cause. So, I've implemented a forced update of one of the "standard" Blender parameters when one of the YafaRay material parameters is modified. It looks like this is solving the problem and effectively updating the material preview. Let's hope this does not cause other problems (?)