New Photon Mapping saving/loading functionality, using Boost serialization library
As requested in http://www.yafaray.org/node/460 the ability to save and load photon maps would be very useful for animations where only the camera moves (fly-through or camera rotating around an object).

I've implemented, with the help of the Boost serialization library, the functionality to "generate only" (traditional photon map way) or "generate and save" as well as "load"

WARNING: When loading Photon Map files, the User is responsible to ensure they match the scene. If the User loads inadequate photon maps, the render results could be totally wrong. USE WITH CARE.