World Background: SmartIBL
As requested in http://yafaray.org/node/566 and http://yafaray.org/node/727  When using IBL we typically suffer from fireflies caused by random sampling getting sometimes samples from the (very bright) sun. It has been suggested to use the SmartIBL technique, blurring the World Texture to reduce fireflies. However that also blurs reflections, which is a no-no.

I've implemented SmartIBL by adding a new Blur parameter to the World IBL. This only works with Image Textures. This blur parameter only affects the lighting, but not the reflections, etc.

The higher the value, the less noise. However higher values will also cause slowdown in the initial blurring process and the scene lighting will not be as realistic. So, use with care!